requirements.txt
python-telegram-bot==20.8
gspread
oauth2client
pytz
